refactor: Abstract away zarr from rkns/rkns.py
refactor: Abstracted awaay zarr.Array

refactor: Abstracted away zarr.Group

refactor: got rid of zarr cals in utils/misc.py,
moved to utils/zarr_utils.py

fix(minor): naming of function override

refactor: Abstract away zarr
