+++ rkns/rkns.py
@@ -5,15 +5,9 @@ import logging
 import warnings
 from pathlib import Path
 from types import EllipsisType
-from typing import TYPE_CHECKING, Iterable, Optional, OrderedDict, cast
+from typing import TYPE_CHECKING, Any, Iterable, Optional, OrderedDict, cast
 
 import numpy as np
-import zarr
-import zarr.core
-import zarr.core.common
-import zarr.core.group
-import zarr.errors
-import zarr.storage
 
 from rkns.adapters.base import RKNSBaseAdapter
 from rkns.adapters.registry import AdapterRegistry
@@ -26,11 +20,10 @@ from rkns.util import (
     RKNSParseError,
     apply_check_open_to_all_methods,
     check_validity,
-    copy_group_recursive,
     get_freq_group,
-    get_or_create_target_store,
 )
-from rkns.util.zarr_util import deep_compare_groups
+from rkns.util.types import JSON
+from rkns.util.zarr_util import ZarrArray
 from rkns.version import __version__
 
 logger = logging.getLogger(__name__)
@@ -39,10 +32,7 @@ logger = logging.getLogger(__name__)
 if TYPE_CHECKING:
     from typing import Self
 
-    from zarr.abc.store import Store
-    from zarr.storage import StoreLike
-
-    from rkns.util import TreeRepr
+    from rkns.util.zarr_util import TreeRepr
 
 
 @apply_check_open_to_all_methods
@@ -57,15 +47,15 @@ class RKNS:
         self.adapter = adapter
 
     @property
-    def patient_info(self) -> zarr.core.common.JSON:
+    def patient_info(self) -> JSON:
         return self.handler.rkns.attrs["patient_info"]
 
     @property
-    def admin_info(self) -> zarr.core.common.JSON:
+    def admin_info(self) -> JSON:
         return self.handler.rkns.attrs["admin_info"]
 
     @property
-    def channel_info(self) -> zarr.core.common.JSON:
+    def channel_info(self) -> JSON:
         return self.handler.rkns.attrs["channel_info"]
 
     def get_channel_names(self) -> list[str]:
@@ -270,10 +260,10 @@ class RKNS:
         )
         return l_signal
 
-    def _get_digital_signal_by_fg(self, frequency_group: str) -> zarr.Array:
+    def _get_digital_signal_by_fg(self, frequency_group: str) -> ZarrArray:
         return self.handler.signals[frequency_group][RKNSNodeNames.rkns_signal.value]  # type: ignore
 
-    def _pminmax_dminmax_by_fg(self, frequency_group: str) -> zarr.Array:
+    def _pminmax_dminmax_by_fg(self, frequency_group: str) -> ZarrArray:
         return self.handler.signals[frequency_group][
             RKNSNodeNames.rkns_signal_minmaxs.value
         ]  # type: ignore
@@ -292,15 +282,14 @@ class RKNS:
         compare_values: bool = True,
         compare_attributes: bool = True,
     ) -> bool:
-        return deep_compare_groups(
-            self.handler.root,
-            other.handler.root,
+        return self.handler.deep_compare(
+            other.handler,
             max_depth=max_depth,
             compare_values=compare_values,
             compare_attributes=compare_attributes,
         )
 
-    def export(self, path_or_store: Store | Path | str) -> None:
+    def export(self, path_or_store: Any | Path | str) -> None:
         """
         Export the RKNS object to a new store, creating a deep copy of all data.
         NOTE: This is a temporary solution, as in the future zarr.copy_all should be available.
@@ -321,18 +310,8 @@ class RKNS:
         This creates a complete copy of the entire RKNS structure, including
         all arrays, groups and their metadata.
         """
-        target_store = get_or_create_target_store(path_or_store)
 
-        if isinstance(target_store, zarr.storage.ZipStore):
-            # The current Zarr implementation has issues with ZIP exports...
-            # That is, attributes are simply not written.
-            raise NotImplementedError()
-
-        try:
-            target_root = zarr.group(store=target_store, overwrite=True)
-            copy_group_recursive(self.handler.root, target_root)
-        finally:
-            target_store.close()
+        self.handler.export_to_path_or_store(path_or_store)
 
     def get_fileformat_of_raw_signal(self) -> FileFormat:
         _raw_signal = self.handler.raw[RKNSNodeNames.raw_signal.value]
@@ -342,9 +321,9 @@ class RKNS:
     @classmethod
     def from_file(
         cls,
-        file_path: StoreLike,
+        file_path: Any,
         populate_from_raw: bool = True,
-        target_store: StoreLike | None = None,
+        target_store: Any | None = None,
     ) -> "RKNS":
         """
         Convenience method to create an RKNS instance from a file.
@@ -410,7 +389,7 @@ class RKNS:
 
 
 class RKNSBuilder:
-    def __init__(self, store: StoreLike | None = None):
+    def __init__(self, store: Any | None = None):
         self._handler = StoreHandler(store)
 
     def _init_base_structure(self) -> None:
@@ -448,7 +427,7 @@ class RKNSBuilder:
 
     def from_file(
         self,
-        file_path: StoreLike,
+        file_path: Any,
         populate_from_raw: bool = True,
     ) -> RKNS:
         """
@@ -497,7 +476,7 @@ class RKNSBuilder:
 
     def from_existing_rkns_store(
         self,
-        store: zarr.storage.StoreLike,
+        store: Any,
         validate: bool = True,
     ) -> RKNS:
         """
