+++ rkns/util/__init__.py
@@ -1,9 +1,7 @@
 from .misc import (
-    TreeRepr,
     apply_check_open_to_all_methods,
     cached_import,
     check_open,
-    group_tree_with_attrs_async,
     import_from_string,
 )
 from .rkns_util import (
@@ -14,7 +12,6 @@ from .rkns_util import (
 )
 from .zarr_util import (
     RKNSParseError,
-    ZarrMode,
     add_child_array,
     copy_attributes,
     copy_group_recursive,
@@ -25,7 +22,6 @@ from .zarr_util import (
 __all__ = [
     "cached_import",
     "import_from_string",
-    "ZarrMode",
     "RKNSNodeNames",
     "check_rkns_validity",
     "add_child_array",
@@ -35,8 +31,6 @@ __all__ = [
     "RKNSParseError",
     "check_open",
     "apply_check_open_to_all_methods",
-    "TreeRepr",
-    "group_tree_with_attrs_async",
     "deep_compare_groups",
     "check_validity",
     "get_freq_group",
